requirements.txt
pypdf
pypdf2
pdf2image 
PyMuPDF
pillow